requirements.txt
discord.py
discord.ui
requests
pytz
PyNaCl
lxml
python-dotenv
nextcord
colorlog
